Implement toggling erase mode with key E
Previous implementation was problematic as it would allow user to enable
erase mode while typing inside of the textbox. Fix is a bit complicated,
but it works. Essentially, the erase shortcut follows the same rules as
the rest of the commands for the color grid; they must not occur if an
event has been propagated by the dom interface controller objects